+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed i18next-react integration with capture models
 - Fixed layout styling with capture model editor
 
+### Removed
+
+- Removed redundant "Back to resource" button from crowdsourcing review page
+
 ## [2.0.0](https://github.com/digirati-co-uk/madoc-platform/releases/tag/v2.0.0) - 2021-XX-XX
 First stable release of Madoc 2.0
 Check out the [documentation](https://docs.madoc.io)